Adding Joanna's posts
I have made some small edits to improve clarity.

Do we want to post 'as is' or should we include a header introducing the post? I can write a brief intro as to how we got this post - or maybe a footer saying where it came from would be better?