Add change clause
The Policy Committee previous approved addition of the clause:

"Any changes to the meaning of the Code of Conduct must be approved by majority vote of both the Policy subcommittee and The Carpentries Executive Council. The Carpentry Community will be informed of these changes and any concerns that are raised will be discussed by the Policy committee."

This PR implements this change and updates language to refer to The Carpentries.